feat: add energy helper module
