+++ src/scripts/post_train.py
@@ -23,7 +23,7 @@ from torch.optim.lr_scheduler import StepLR
 from torch.utils.data import DataLoader
 
 from src.file_config import FLOWBACK_BASE, FLOWBACK_DATA, FLOWBACK_JOBDIR
-from src.utils.energy import ensure_charmm_ff
+from src.utils.energy_helpers import ensure_charmm_ff
 from src.adjoint import (
     adjoint_matching_loss,
     sigma,
