+++ src/utils/energy.py
@@ -3,220 +3,18 @@ from rdkit.Chem import AllChem
 import numpy as np
 import mdtraj as md
 import torch
-from rdkit.Chem.rdMolDescriptors import CalcMolFormula
-from rdkit import RDLogger
 from rdkit import rdBase
 from openmm import *
 import os
 from openmm.app import *
 from openmm.unit import *
-import argparse
 import tempfile
-import MDAnalysis as mda
 import subprocess
+from .energy_helpers import compute_all_distances, ensure_charmm_ff, _map_original_to_processed_indices, silence_atoms_and_shift_charge
 import warnings
-import gc
-from collections import defaultdict
-import io
-import time
-from pathlib import Path
-import shutil
-import re
-from src.file_config import FLOWBACK_BASE, FLOWBACK_FF
-from typing import Dict, List, Tuple, Optional
 
 warnings.filterwarnings('ignore')
 
-def _osremove(f):
-    try:
-        os.remove(f)
-    except FileNotFoundError:
-        pass
-
-def compute_all_distances(traj):
-    idxs = np.arange(traj.top.n_atoms)
-    grid = np.array(np.meshgrid(idxs, idxs)).T.reshape(-1, 2)
-    pairs = grid[grid[:, 0] > grid[:, 1]]
-    dists = md.compute_distances(traj, pairs)
-    return dists
-
-
-
-
-_TOPLEVEL_SUBSECTIONS = {
-    "replace", "add", "delete"
-}
-
-_BLOCK_HEADER_RE = re.compile(r'^\[\s*([^\]]+?)\s*\]\s*$', re.MULTILINE)
-
-def _is_subsection(name: str) -> bool:
-    # Case-insensitive match for subsection keywords
-    return name.strip().lower() in _TOPLEVEL_SUBSECTIONS
-
-def _find_first_existing(files: List[Path], preferred_names: Tuple[str, ...]) -> List[Path]:
-    """Return files in priority order: any explicitly preferred names first (if present),
-    else all matching files as given."""
-    preferred = []
-    others = []
-    for f in files:
-        if f.name in preferred_names:
-            preferred.append(f)
-        else:
-            others.append(f)
-    # Try preferred ones first, but still consider others for missed entries
-    return preferred + others
-
-def _gather_ff_tdb_files(ff_dir: Path, suffix: str) -> List[Path]:
-    """Gather *.n.tdb or *.c.tdb files from a force field directory."""
-    return sorted(ff_dir.glob(f"*{suffix}.tdb"))
-
-def _extract_named_blocks_from_text(text: str, wanted_names: List[str]) -> Dict[str, str]:
-    """
-    Extract blocks by name from a .tdb file's text.
-    A block starts at a top-level header line: [ NAME ] where NAME is not a lowercase subsection.
-    The block ends just before the next top-level block header.
-    """
-    results: Dict[str, str] = {}
-    # Find all bracket headers with their positions
-    matches = list(_BLOCK_HEADER_RE.finditer(text))
-    # Precompute header names and which are top-level vs subsection
-    headers = []
-    for m in matches:
-        name = m.group(1).strip()
-        headers.append((name, m.start(), m.end(), not _is_subsection(name)))
-
-    # Iterate top-level headers and grab content until next top-level header
-    for i, (name, start, end, is_top) in enumerate(headers):
-        if not is_top:
-            continue
-        block_start = start
-        # Find next top-level header's start
-        j = i + 1
-        block_end = len(text)
-        while j < len(headers):
-            n2, s2, e2, is_top2 = headers[j]
-            if is_top2:
-                block_end = s2
-                break
-            j += 1
-        block_text = text[block_start:block_end].rstrip() + "\n"
-        # If this top-level name is one we want, and we don't already have it, store it
-        if name in wanted_names and name not in results:
-            results[name] = block_text
-
-    return results
-
-def _extract_blocks(ff_dir: Path,
-                    tdb_suffix: str,
-                    names_to_find: List[str],
-                    preferred_files: Tuple[str, ...]) -> Dict[str, str]:
-    """
-    Search across all matching *.tdb files (prioritizing names in preferred_files)
-    and extract the given block names.
-    """
-    candidates = _gather_ff_tdb_files(ff_dir, tdb_suffix)
-    if not candidates:
-        raise RuntimeError(f"No '*{tdb_suffix}.tdb' files found in '{ff_dir}'")
-
-    # Prioritize common merged filenames first (e.g., merged.n.tdb / merged.c.tdb)
-    ordered_files = _find_first_existing(candidates, preferred_files)
-
-    found: Dict[str, str] = {}
-    remaining = set(names_to_find)
-
-    for f in ordered_files:
-        if not remaining:
-            break
-        text = f.read_text()
-        blocks = _extract_named_blocks_from_text(text, list(remaining))
-        for k, v in blocks.items():
-            found[k] = v
-        remaining -= set(blocks.keys())
-
-    if remaining:
-        searched = ", ".join(p.name for p in ordered_files)
-        missing = ", ".join(sorted(remaining))
-        raise RuntimeError(
-            f"Missing required terminal entries [{missing}] in {tdb_suffix} databases. "
-            f"Searched files: {searched}"
-        )
-    return found
-
-def ensure_charmm_ff(version: str = 'auto') -> Path:
-    """Copy user's CHARMM force field into FLOWBACK_FF with merged termini
-    extracted directly from the force field's own terminal databases.
-    """
-    if version == 'auto' or version == 'charmm':
-        keyword = 'charmm'
-    else:
-        keyword = version
-
-    dest = Path(FLOWBACK_FF) / f"{keyword}.ff"
-    if dest.exists():
-        return dest
-
-    gmxlibrary = os.environ.get("GMXLIB")
-    if not gmxlibrary:
-        raise RuntimeError("$GMXLIB is not set; cannot locate CHARMM force field")
-
-    entries = [e for e in os.listdir(gmxlibrary)
-               if e.lower().startswith(keyword) and e.endswith(".ff")]
-    if not entries:
-        raise RuntimeError(f"No CHARMM force field found in GMXLIB '{gmxlibrary}'")
-
-    def version_key(name: str) -> int:
-        m = re.search(r"(\d+)", name)
-        return int(m.group(1)) if m else 0
-
-    entries.sort(key=version_key)
-    src = Path(gmxlibrary) / entries[-1]
-
-    # Copy the entire force field directory first
-    shutil.copytree(src, dest, dirs_exist_ok=True)
-
-    # Extract needed terminal blocks from the *source* FF (before we delete in dest)
-    # Prioritize merged files if present; still search others as fallback.
-    # N-termini: exact names
-    n_names = ["PRO-NH2+", "GLY-NH3+", "NH3+"]
-    n_blocks = _extract_blocks(
-        ff_dir=src,
-        tdb_suffix=".n",
-        names_to_find=n_names,
-        preferred_files=("merged.n.tdb",)
-    )
-
-    # C-terminus: prefer COO-. Some CHARMM packs may name the default negative terminus "CTER".
-    c_targets_ordered = ["COO-"]  # search order; we will pick the first found
-    c_blocks_all = _extract_blocks(
-        ff_dir=src,
-        tdb_suffix=".c",
-        names_to_find=c_targets_ordered,
-        preferred_files=("merged.c.tdb",)
-    )
-    # Choose COO- if present; otherwise use CTER
-    c_choice_name = "COO-" if "COO-" in c_blocks_all else "CTER"
-    c_block_text = c_blocks_all[c_choice_name]
-
-    # Now remove existing terminal DBs in the *destination* and write merged ones
-    for pattern in ("*.n.tdb", "*.c.tdb"):
-        for f in dest.glob(pattern):
-            try:
-                f.unlink()
-            except FileNotFoundError:
-                pass
-
-    # Write merged.n.tdb with the three N-terminus entries in a fixed order
-    merged_n_path = dest / "merged.n.tdb"
-    merged_n_text = ""
-    for name in n_names:
-        merged_n_text += n_blocks[name].rstrip() + "\n\n"
-    merged_n_path.write_text(merged_n_text.rstrip() + "\n")
-
-    # Write merged.c.tdb with the chosen C-terminus entry
-    merged_c_path = dest / "merged.c.tdb"
-    merged_c_path.write_text(c_block_text)
-
-    return dest
 
 
 
@@ -432,118 +230,4 @@ def rdkit_structure_to_energy(topology: md.Topology, xyz: np.ndarray):
     return energy, gradients.reshape(-1, 3)
 
 # @time_elapsed
-def _map_original_to_processed_indices(original_pdb, processed_pdb):
-    """
-    Create a mapping from original PDB atom indices to processed PDB indices after pdb2gmx.
-    
-    Parameters:
-    -----------
-    original_pdb : str
-        Path to the original PDB file
-    processed_pdb : str
-        Path to the processed PDB file (after pdb2gmx)
-        
-    Returns:
-    --------
-    list
-        List where index position is the original atom index and value is the new index.
-        -1 indicates no match was found for that original atom.
-    """
-    # Load both structures
-    orig_u = mda.Universe(original_pdb)
-    proc_u = mda.Universe(processed_pdb)
-    
-    # Initialize mapping list with -1 (indicating no match)
-    n_orig_atoms = len(orig_u.atoms)
-    index_map = -1 * np.ones(n_orig_atoms).astype(int)
-    
-    # Create unique identifiers for each atom in processed structure
-    proc_dict = {}
-    for atom in proc_u.atoms:
-        identifier = (atom.resnum, atom.resname, atom.name)
-        proc_dict[identifier] = atom.ix
-    
-    # Map original atoms to processed atoms
-    for atom in orig_u.atoms:
-        identifier = (atom.resnum, atom.resname, atom.name)
-        if identifier in proc_dict:
-            index_map[atom.ix] = proc_dict[identifier]
-        elif atom.name == 'O':
-            identifier = (atom.resnum, atom.resname, 'OT1')
-            index_map[atom.ix] = proc_dict[identifier]
-        elif atom.resname == 'ILE' and atom.name == 'CD1':
-            identifier = (atom.resnum, atom.resname, 'CD')
-            index_map[atom.ix] = proc_dict[identifier]
-    return index_map
 
-def silence_atoms_and_shift_charge(nbforce, topology, mute, context=None):
-    """
-    Zero LJ and charge on each atom in *mute* and transfer that charge
-    to its first bonded neighbour that is **not** in *mute*.
-
-    Parameters
-    ----------
-    nbforce : openmm.NonbondedForce
-        The force we are modifying.
-    topology : openmm.app.Topology
-        Needed to discover bonding so we know where to move the charge.
-    mute : iterable of int
-        Particle indices to silence.
-    context : openmm.Context, optional
-        Pass the existing Context if you already created one so we can
-        push the new parameters with `updateParametersInContext`.
-    """
-    mute    = set(int(i) for i in mute)
-    shifts  = defaultdict(lambda: 0.0*elementary_charge)   # charge → heavy atom
-    nbforce.setUseDispersionCorrection(False)
-    # nbforce.setNonbondedMethod(NonbondedForce.CutoffPeriodic)
-    # ------------------------------------------------------------------
-    # 1. build a quick neighbour map from the Topology
-    # ------------------------------------------------------------------
-    neighbours = defaultdict(list)
-    for bond in topology.bonds:
-        i, j = bond[0].index, bond[1].index
-        neighbours[i].append(j)
-        neighbours[j].append(i)
-
-    # ------------------------------------------------------------------
-    # 2. zero LJ & charge on the muted atoms, remember the charge removed
-    # ------------------------------------------------------------------
-    for idx in mute:
-        q, sigma, eps = nbforce.getParticleParameters(idx)
-
-        # accumulate the charge so we can add it to the parent later
-        if abs(q.value_in_unit(elementary_charge)) > 1e-12:
-            # first bonded neighbour that is NOT muted
-            try:
-                parent = next(n for n in neighbours[idx] if n not in mute)
-                shifts[parent] += q
-            except StopIteration:
-                warnings.warn(f"Atom {idx} is muted but has no non-muted neighbours;"
-                              " total charge will not be conserved!")
-
-        # zero LJ *and* charge
-        nbforce.setParticleParameters(idx,
-                                      0.0*elementary_charge,
-                                      sigma,
-                                      0.0*kilojoule_per_mole)
-
-    # ------------------------------------------------------------------
-    # 3. apply the accumulated charge shifts to the parent atoms
-    # ------------------------------------------------------------------
-    for idx, dq in shifts.items():
-        q, sigma, eps = nbforce.getParticleParameters(idx)
-        nbforce.setParticleParameters(idx, q + dq, sigma, eps)
-
-    # ------------------------------------------------------------------
-    # 4. update every exception so chargeProd matches the new charges
-    # ------------------------------------------------------------------
-    for k in range(nbforce.getNumExceptions()):
-        i, j, qprod, sigma, eps = nbforce.getExceptionParameters(k)
-        
-        qi = nbforce.getParticleParameters(i)[0]
-        qj = nbforce.getParticleParameters(j)[0]
-        if topology.atom(i).element.name == 'hydrogen' or topology.atom(j).element.name =='hydrogen':
-            nbforce.setExceptionParameters(k, i, j, qi*qj, sigma, 0.0*kilojoule_per_mole)
-        else:
-            nbforce.setExceptionParameters(k, i, j, qi*qj, sigma, eps)
